ccgp_bidding/requirement.txt
requests==2.9.1
pykafka==2.3.1
pika==0.10.0
pymongo
cassandra-driver
psycopg2=2.4.5
tornado
boto3
